不要なクラスを削除
